refactor: организовать код
- разделить код на feature based файловую структуру
- и упростить тем самым импорты
- вынести декораторы API
