+++ src/common/decorators/controller/index.ts
@@ -1,1 +1,3 @@
 export * from './find-all';
+export * from '../../offset-pagination/offset-pagination-queries';
+export * from './cursor-paginated-request';
