+++ src/common/features/filtering/index.ts
@@ -1,0 +1,2 @@
+export * from './FilteringParams';
+export * from './filtering-query.decorator';
