+++ src/common/features/sorting/index.ts
@@ -1,0 +1,2 @@
+export * from './SortingParams';
+export * from './sorting-query.decorator';
