+++ src/common/offset-pagination/constants.ts
@@ -1,2 +1,4 @@
 export const MAX_LIMIT = 100;
 export const MIN_LIMIT = 1;
+export const DEFAULT_LIMIT = 10;
+export const DEFAULT_PAGE = 1;
