+++ src/database/functions/index.ts
@@ -1,0 +1,2 @@
+export * from './clearDatabase';
+export * from './seedDatabase';
