+++ src/database/seeds/index.ts
@@ -1,0 +1,1 @@
+export * from './survey.seeder';
